+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show the comment like counts in the UI. (mddburgess/winston#88)
 - Show error messages on the pull channel modal. (mddburgess/winston#91)
 - Show channel video, view, and subscriber counts on channel cards. (mddburgess/winston#91)
+- Add button on channel details page to refresh channel data. (mddburgess/winston#92)
+- Show channel statistics on channel details page. (mddburgess/winston#94)
 
 ### Changed